buffys roleplay chat with typing indicator
Mod that replaces both Buffy's roleplay chat and Nater's Chat Bubble. It will put a "Typing..." message over the heads of anyone typing, but not save it to the client chat log. Works in quick local multiplayer test server.

Known bug: It filters out all messages that contain "Typing..." from displaying in the chat log so "Haha, I'm going to kill you while you are typing..." will show over the speaker's head but not show on the client chat log. It should still show up in the server chat logs just fine. If this is a major concern, I can put an invisible unicode character inside the "Typing..." message for make it harder to false positive, but keeping from doing so for now for code readability.